Bug 1866057 - Fix SourceRepository/SourceStamp Snap upstream build on try r=jcristau
When we perform a try build for a Snap package using upstream packaging,
we are reusing the gecko clone produced for us by run-task. However,
this clone is from mozilla-unified, so we end up with the wrong link in
about:buildconfig.

Also by default Snap upstream package removes the .hg directory for
nightly builds (not others because they download tar.xz sources which
have the sourcestamp.txt) and since it's from a clone we expect to be
able to run hg commands to get:
 - repo address
 - current changeset

By removing the .hg this is explicitely broken. Since the clone process
already relies on the correct repo URL and changeset, we just force
those values in the env so correct about:buildconfig can be shown.

Differential Revision: https://phabricator.services.mozilla.com/D194496

UltraBlame original commit: 473fafa234fdc295dad2914e4c7b49f482f3731a