Bug 1850914 - add script to update test parameters.yml. r=releng-reviewers,ahal,gabriel
If called with no arguments it'll update all files (except try.yml and
mr-onpush-geckoview.yml).  It can also be called with a set of filenames
as arguments and will update just those files.

For on-push and cron tasks, we look up the taskcluster index for the
latest corresponding decision task (or possibly an older one in case of
DONTBUILD pushes).  For release promotion actions, we query the public
shipit API to find the latest matching task.

Differential Revision: https://phabricator.services.mozilla.com/D187137

UltraBlame original commit: 02bdacaadcbbf0badf5918ec7ce2b0d3609fcc64