Bug 1850914 - avoid checking for existing tc indexes when running test_generate_params on CI. r=releng-reviewers,gbrown
The target_tasks methods for ship_geckoview and nightly_desktop return
an empty graph when running on automation if they detect a taskcluster
index from a previous run on the same revision.  This wasn't a problem
with old parameters since the corresponding indexed tasks are expired,
but with updated parameters it starts triggering.  We don't want that
logic here, since we're asserting we get a nonempty graph.

Differential Revision: https://phabricator.services.mozilla.com/D194523

UltraBlame original commit: c8442739bed98c74f40283151ec240c3ec6ac053