Bug 1865625 - force attestation to none in macOS platform passkey support r=jschanck
On macOS, the API for platform passkeys doesn't support attestation, so it must
always be "none". Also, when using a cross-platform device, the API shows no
consent UI for non-none attestation, so this forces that to "none" as well, for
now.

Differential Revision: https://phabricator.services.mozilla.com/D194935

UltraBlame original commit: 03593d473c691a886c594c500071df8a217085ce