Bug 1616524 - [devtools] Revise eager evaluation implementation. r=devtools-reviewers,nchevobbe
There is no need to add all global to the side effect free Debugger.
We should only add the evaluation context globals.
In the typical case it will be the target global.
But for content process targets, it may be multiple globals, which we all need to register.

Also ensure clearing onNativeCall callback to be really sure this is disabled,
even if the Debugger is leaked for some unexpected reason.
Having this listener set may drastically slow down executions by disabling the JIT.
There was no known scenarios where it happened, but it is to be extra safe.

Also simplified a few things in evalWithDebugger as eslint started complaining about its complexity.

Differential Revision: https://phabricator.services.mozilla.com/D195021

UltraBlame original commit: 01a94921a62340d1afeda517d25fc4b29e7a2530