Bug 1814969: Migrate ContextualIdentityService strings to Fluent r=geckoview-reviewers,fluent-reviewers,mconley,owlish,flod
This includes a containers.json version increment and the necessary migration,
as the StringBundle identifiers are replaced with a Fluent reference.

As it's not possible to filter the attributes that are applied with DOM localization,
the code in utilityOverlay.js is a little clumsy. The non-acceskey versions are needed
due to bug 1297738, i.e. the long-press menu available from the tab bar down arrow.

Previously, the "Manage containers" menu item and its accesskey were defined in two places.
These are here combined into one.

Differential Revision: https://phabricator.services.mozilla.com/D194510

UltraBlame original commit: 2579a5934476edbd55f206ec87d13582ec85a5ca