Bug 1868826 - Increase frequency of CaR sp3 tests to be on a per-commit basis. r=perftest-reviewers,sparky
Right now with the current scheduling, speedometer tests for
Chromium-as-Release only run once a day. This patch increases that
frequency to run the test on every push to mozilla-central, so that we
can get a bit more data easily.

Differential Revision: https://phabricator.services.mozilla.com/D195809

UltraBlame original commit: 75e0ede09503eccf1cbef8632f794d1929320b8c