Bug 1867542 - Introduce a generic configuration change interface to the PEM inteface. r=jolin
A configuration change is an array of sum types that each describe a
configuration change. Sometimes the configuration can be the empty type, in
which case the PEMs should attempt to set the value to a default that is
maybe platform-specific.

We will add to this list of possible changes over time but this is a good first
set.

Depends on D195145

Differential Revision: https://phabricator.services.mozilla.com/D195146

UltraBlame original commit: 509a405ae33e1dfb5b68962e8a7170d5872dd5a7