Bug 1868680 - Stop prewarming the ARM AVD. r=geckoview-reviewers,owlish
prewarming the AVDs was done for CI, where AVDs are used only once in
each task using them, and the cost of first use is high. Prewarming
reduces that overhead.

However, the ARM AVD is not used on CI (anymore?), so we don't need
them prewarmed. This unblocks the situation wrt the failure to run
the emulator on CI, while not addressing the underlying issues, so that
CoT keys can be rotated without waiting for the full situation to be
resolved.

Differential Revision: https://phabricator.services.mozilla.com/D195820

UltraBlame original commit: 3ae67b2bcfa9121c2770b546f543737368b59393