Bug 1850149 - Include `previous_os_version` in `default-browser-agent` telemetry pings. r=nrishel
This adds both the current `os_version` and the `previous_os_version`
to both legacy and Glean pings.  This appears to duplice the built-in
Glean `os_version` metric, but they are *not* the same format as the
legacy telemetry: Glean records the UBR as a separate metric.

Differential Revision: https://phabricator.services.mozilla.com/D196120

UltraBlame original commit: 49dddecc64c63ffe48f462e97f0b8eef012f480c