Bug 1868304 - Add temporary code in Nightly to diagnose SetReservedClientInfo assertion. r=pbone,dom-worker-reviewers,asuth
To help investigate the crash in bug 1761208, this patch adds code to
register the top frames of the call stack when reservedClientInfo is
set, such that it can be recovered from crash dumps.

We only add this code for Nightly Windows x64 builds because we get a
Nightly crash on this assertion for almost every Nightly build, always
coming from Windows x64. So this should be enough and there is no need
to impact other builds.

This code only needs to live for a few Nightly builds and can/should be
removed once we have caught the faulty call stack.

Differential Revision: https://phabricator.services.mozilla.com/D195525

UltraBlame original commit: 9140590c2e997acaae28ed63bc8124666cb189ab