Bug 1868336 - Part 1: Move DeduplicationStringHasher definition out of Tenuring.h r=sfink
This class is internal to tenuring and is only used from there, so there's no
reason to put these in the header file.

I suspect this doesn't need to be templated but I left it as it was.

Differential Revision: https://phabricator.services.mozilla.com/D195538

UltraBlame original commit: bf216deaa62691f3e10ce3f6d9f595ae55c2394a