Bug 1868336 - Part 5: Fix pretenuring reporting r=sfink
The AllocSite::hasScript method was broken by the wasm GC changes. This should
return false when the script pointer is null too.

The only thing this actually affects is the output when JS_GC_REPORT_PRETENURE
is set, since we don't otherwise call it when the script pointer is null.

While we're here, add a header line to this output to help understand it.

Differential Revision: https://phabricator.services.mozilla.com/D195765

UltraBlame original commit: fa5f08ef566e72f886920db42ce1806a42717fac