Bug 1869527 - Explicitly unparent old SurfaceControls following GPU process crash. r=geckoview-reviewers,owlish
We used to avoid doing this in order to avoid a blank screen for a
second whilst rendering is reinitialized. However, not doing so is
causing severe rendering glitches on Samsung S22 devices.

The subsequent patch in this series is actually going to disable
SurfaceControl rendering anyway, as bug 1824083 provides an
alternative solution to recovering from GPU process crashes. However,
it is better to fix this issue now anyway in case we decide to
re-enable it one day.

Differential Revision: https://phabricator.services.mozilla.com/D196168

UltraBlame original commit: 00826681d51d47d70a523d2fc99bce9b87a0c66c