Bug 1870100 - Fade the titlebar entirely rather than by changing text color. r=dao,desktop-theme-reviewers
This avoids big restyles when users have a massive number of tabs. A
color change inherits, so we need to propagate the color change to the
whole subtree, which with a lot of tabs can be huge (66k elements in the
profile in the bug).

Instead, don't change the text color on inactive windows, and fade the
titlebar entirely using opacity, which is a one-element change.

Differential Revision: https://phabricator.services.mozilla.com/D196543

UltraBlame original commit: 80d2ca5e4bece1bfb30ea95421c0c410c822b835