Bug 1855301 - Support observing NativeObject property modification r=jandem,iain
Watching some objects for property modification will be key in using fuses to
assert system invariants.

This patch instruments various parts of the code to catch potential writes to
regular objects slots which corresponding to string or symbol named properties.
It does not attempt to cover elements, which will have to be handled
separately.

This patch focuses only on *script exposed* objects, which means that it does
not attempt to support for watching anonymous slots, such as those that exist
in EnvironmentObjects or ModuleObjects.

The set of watches was built through an analysis of `NativeObject::setSlot`
calls.

Differential Revision: https://phabricator.services.mozilla.com/D195261

UltraBlame original commit: 6f96fe49c54e13ee062e2bac2d0d539dc3b489a6