Bug 1870911: Don't inadvertently clobber masonry bits just because a nested grid can't be a subgrid. r=TYLin
This ComputeSelfSubgridMasonryBits() function has some logic to determine
whether a grid is using subgrid layout (i.e. participating in a parent grid)
and also whether it's using masonry layout at all.

There's one early-return for a case where we can't be a subgrid, which
inadvertently also disregards the masonry state that we've set. It looks like
this was an oversight, since in fact we can still use masonry regardless of
whether we're allowed to be a subgrid.

Differential Revision: https://phabricator.services.mozilla.com/D196891

UltraBlame original commit: 582f5608c0a3d1f7fa784427a517c3923c18cd0f