Bug 1869942 - Temporarily pass clicks on inacessible richlistbox treecol in a11y-checks while their accessibility is being fixed. r=Jamie,settings-reviewers,Gijs
Add a temporary check in AccessibilityUtils to pass clicks on inacessible richlistbox treecol so we could remove the `fail-if` notations from the affected test manifests until the bug 1848397 is being resolved. This would allow a11y-checks to continue monitor these affected tests and to avoid unknowingly introducing new inaccessible controls because of that.

When the bug 1848397 is fixed, remove the temporary check introduced in this patch.

Differential Revision: https://phabricator.services.mozilla.com/D196376

UltraBlame original commit: 0068dcbf07fd49f953a4c4780abb50c90c5061fa