Bug 1869835 - Part 2: Do some code deduplication and convert UserTimingMarker to the new schema. r=mstange,profiler-reviewers
This will allow ETW to collect the additional payload on the UserTiming markers. As well as allow the provider to filter on these markers in order to minimize overhead from other markers that don't need to be collected.

Differential Revision: https://phabricator.services.mozilla.com/D196332

UltraBlame original commit: 2154f1cbb973bfc5c9937bb2fd61c6f02990baba