Bug 1871540 - Fix an inspector assertion when inspecting ::cue pseudos. r=dshin
I noticed that we weren't finding the :host::cue rule in the inspector,
even though it was applying and working.

When we're a NAC pseudo-element, also look at the relevant shadow root.

Could probably write a test for this, but I think I want to make the
setup for ::cue (which is the only weird pseudo like this) a bit nicer,
so not sure it's worth it.

Differential Revision: https://phabricator.services.mozilla.com/D197185

UltraBlame original commit: 7d1bc9b657a7af139393384d647209e108abdd2f