Bug 1854010 - Restrict fixed/sticky hit test info to items fixed to the root. r=hiro,botond
Fixed and Sticky position side bits should only be necessary for fixed
or sticky position items that are fixed or stuck to the root content. As
a result, there is no need to add the fixed or sticky position hit test
info to items that are not.

Differential Revision: https://phabricator.services.mozilla.com/D192272

UltraBlame original commit: dd6cba997d6a64a1e7e180ad1a2d1feecb895708