Bug 1833482 part 0: Rename CollectSubgridForAxis to CollectSubgridItemsForAxisRecursive. r=emilio,layout-reviewers
This patch doesn't change behavior; it's just a rename.

We have an "entry-point" function CollectSubgridItemsForAxis that the outer
grid calls, and most of its work is done by a recursive helper function.

Before this patch, the recursive helper function has nearly the same name but
with the word "Items" removed, which doesn't make any sense, since it's still
collecting items from subgrids.

Let's just rename the recursive helper to match its entry-point, with a
"Helper" suffix, and document that it's a recursive helper. This makes the
function name a little longer, but not too much longer.

(I'm bothering to do this because the next patch in this series will add a
similar pair of functions, and I want the new functions to use the same naming
scheme as these existing functions.)

Differential Revision: https://phabricator.services.mozilla.com/D197232

UltraBlame original commit: 75f5d477db6fa01cf8884fce671a8a306d24ad82