Bug 1872505 - Use NSString in InsertText* paths. r=m_kato
The InsertText* code paths required an NSAttributedString instance but
only ever used their NSString representation. Callers nearly always have
an NSString, so this results in a lot of unnecessary NSAttributedString
allocations to satisfy this interface.

The primary caller is [ChildView insertText:replacementRange:], which
can receive either an NSString or an NSAttributedString string from the
NSTextInputClient protocol, but this is usually (always?) an NSString in
practice.

We can instead change these methods to receive NSString, simplifying the
callers and removing a bunch of short-lived NSAttributedString objects.

Differential Revision: https://phabricator.services.mozilla.com/D197451

UltraBlame original commit: 89fae03510c9900abefe1653e2f0b380dbcb616e